moved new tests to a dedicated file
